Revert "Add files via upload"
This reverts commit 3286196f0f2dc25526d162ea9a0b88cdbdb4c837.
