+++ fytProject/app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "FYTAPP",
-    "slug": "FYTAPP",
+    "name": "fytProject",
+    "slug": "fytProject",
     "privacy": "public",
     "sdkVersion": "31.0.0",
     "platforms": [
@@ -10,9 +10,7 @@
     ],
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./assets/images/icon.png",
     "splash": {
-      "image": "./assets/images/splash.png",
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
