+++ fytProject/package.json
@@ -8,16 +8,12 @@
     "test": "jest"
   },
   "dependencies": {
-    "@expo/samples": "^2.1.1",
     "@expo/vector-icons": "^8.0.0",
     "expo": "^31.0.2",
-    "native-base": "^2.8.1",
-    "package.json": "^2.0.1",
+    "native-base": "2.8.1",
     "react": "^16.6.3",
     "react-native": "^0.57.8",
-    "react-native-material-ui": "^1.30.1",
-    "react-native-swipe-gestures": "^1.0.3",
-    "react-navigation": "^1.6.1",
+    "react-navigation": "1.5.1",
     "tcomb-form-native": "^0.6.20"
   },
   "devDependencies": {
