--- fytProject/packagebackup.json
@@ -1,32 +1,0 @@
-{
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "eject": "expo eject",
-    "test": "node ./node_modules/jest/bin/jest.js --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
-  },
-  "dependencies": {
-    "@expo/samples": "2.1.1",
-    "@material-ui/core": "^3.7.0",
-    "expo": "^31.0.6",
-    "package.json": "^2.0.1",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-native": "^0.57.8",
-    "react-native-material-ui": "^1.30.1",
-    "react-native-vector-icons": "^4.0.0",
-    "react-navigation": "^2.18.2",
-    "rect": "^1.2.1",
-    "typeface-roboto": "0.0.54"
-  },
-  "devDependencies": {
-    "babel-preset-expo": "^5.0.0",
-    "jest-expo": "^31.0.0"
-  },
-  "private": true
-}
